+++ Tests/UITests/TestApp/ScheduleView.swift
@@ -11,62 +11,7 @@ import SpeziViews
 import SwiftUI
 
 
-struct QuestionnaireEventDetailView: View {
-    private let event: Event
-
-    @Environment(\.dismiss)
-    private var dismiss
-
-    var body: some View {
-        NavigationStack {
-            List {
-                DefaultTileHeader(event, alignment: .center)
-
-                Section {
-                    Text(event.task.instructions)
-                } header: {
-                    Text("Instructions")
-                        .detailHeader()
-                }
-
-                if let about = event.task.about {
-                    Section {
-                        Text(LocalizedStringResource(about))
-                    } header: {
-                        Text("About")
-                            .detailHeader()
-                    }
-                }
-            }
-                .navigationTitle("More Information")
-                .navigationBarTitleDisplayMode(.inline)
-                .toolbar {
-                    Button("Close") {
-                        dismiss()
-                    }
-                }
-        }
-    }
-
-    init(_ event: Event) {
-        self.event = event
-    }
-}
-
-
-extension Text {
-    func detailHeader() -> some View {
-        self
-            .font(.title3)
-            .fontWeight(.bold)
-            .textCase(.none)
-            .foregroundStyle(.primary)
-    }
-}
-
-
 struct ScheduleView: View {
-    // TODO: show today vs tomorrow?
     @EventQuery(in: Date.today..<Date.tomorrow)
     private var events
 
@@ -85,8 +30,8 @@ struct ScheduleView: View {
                     )
                 } else {
                     // TODO: reusable list!
+                    // TODO: today and tomorrow headings!?
                     List {
-                        // TODO: today and tomorrow headings!?
                         Section {
                             Text("Today")
                                 .foregroundStyle(.secondary)
@@ -99,15 +44,11 @@ struct ScheduleView: View {
 
 
                         ForEach(events) { event in
-                            // TODO: snapshot tests with different alignments + info button + with/without category
                             Section {
-                                // TODO: completed doesn't work!
                                 InstructionsTile(event) {
                                     QuestionnaireEventDetailView(event)
                                 } action: {
-                                    event.complete() // TODO: complete with keys?
-
-                                    print("Outcome is now \(event.outcome)")
+                                    event.complete()
                                 }
                             }
                         }
